UTH-35: hide contracts tab from create applicant modal
